Fix test compilation errors failing Docker build

+++ backend/src/test/java/com/luxurystay/service/HotelServiceImplTest.java
@@ -55,7 +55,7 @@ public class HotelServiceImplTest {
                 .name("Grand Plaza")
                 .city("New York")
                 .country("USA")
-                .rating(4.5)
+                .rating(BigDecimal.valueOf(4.5))
                 .build();
     }
